+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]: https://github.com/trussed-dev/ctap-types/compare/0.6.0-rc.1...HEAD
 
 - Rename `authenticator_config` to `config`.
+- Add `platform-serde` feature for additional `Serialize` and `Deserialize` implementations not required by authenticators.
 
 ## [0.6.0-rc.1] 2026-05-21